requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
cloudinary==1.44.0
Django==5.1.6
django-cloudinary-storage==0.3.0
django-debug-toolbar==5.0.1
django-htmx==1.23.0
django-jazzmin==3.0.1
django-shortuuidfield==0.1.3
django-taggit==6.1.0
Django-Verify-Email==3.0.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.1.0
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.4
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
